+++ app/src/main/java/com/gps19/app/GpsStatusManager.kt
@@ -17,12 +17,12 @@ class GpsStatusManager @Inject constructor(
 ) {
     /**
      * GpsStatusManager: Centralized reactive Flow for the GPS-Index.
+     * v8.9.79: Issue #014 - Type Migration: Standardized accuracy to Double.
      * v8.8.21:
      * - Modularization: Direct integration with TelemetryUtils for signal scoring.
-     * v8.8.28: Resolved unresolved reference to isValidLocation.
      */
     fun observeGpsIndex(nowFlow: Flow<Long>): Flow<GpsIndexData> {
-        data class IndexParams(val gpsTs: Long, val maxAccuracy: Float, val satsUsed: Int)
+        data class IndexParams(val gpsTs: Long, val maxAccuracy: Double, val satsUsed: Int)
 
         return combine(
             settingsRepository.appModeFlow,
@@ -41,7 +41,7 @@ class GpsStatusManager @Inject constructor(
             } else null
             
             params to now
-        }.scan(GpsIndexData(0f, 0f, 0f, 0f) to (null as IndexParams?)) { state, (params, now) ->
+        }.scan(GpsIndexData(0.0, 0.0, 0.0, 0.0) to (null as IndexParams?)) { state, (params, now) ->
             val lastParams = state.second
             
             // v5.941 Monotonicity & Heartbeat Firewall
@@ -59,7 +59,7 @@ class GpsStatusManager @Inject constructor(
                 )
                 index to activeParams
             } else {
-                GpsIndexData(0f, 0f, 0f, 0f) to null
+                GpsIndexData(0.0, 0.0, 0.0, 0.0) to null
             }
         }.map { it.first }
     }
